+++ R/02-gimap_filter.R
@@ -45,7 +45,7 @@
 #' timepoint replicates
 #' @export
 #' @import dplyr
-#' @examples \donttest{
+#' @examples \dontrun{
 #'
 #' gimap_dataset <- get_example_data("gimap", data_dir = tempdir()) %>%
 #'   gimap_filter()
@@ -221,7 +221,7 @@ gimap_filter <- function(.data = NULL,
 #' count zero for at least one sample/time point and a report df `reportdf` for
 #' the number and percent of pgRNA which have a count zero for at least one
 #' sample/time point
-#' @examples \donttest{
+#' @examples \dontrun{
 #' gimap_dataset <- get_example_data("gimap", data_dir = tempdir())
 #' qc_filter_zerocounts(gimap_dataset)
 #'
@@ -286,7 +286,7 @@ qc_filter_zerocounts <- function(gimap_dataset,
 #' low plasmid log2 CPM (column of interest is `plasmid_cpm_filter`) and a
 #' report df `reportdf` for the number and percent of pgRNA which have a low
 #' plasmid log2 CPM
-#' @examples \donttest{
+#' @examples \dontrun{
 #' gimap_dataset <- get_example_data("gimap", data_dir = tempdir())
 #'
 #' qc_filter_plasmid(gimap_dataset)